Revert unrelated change
